# Update Marshal to use strings.Builder